refactors PostPreview (previously ImageContainer)
