--- src/components/image/_image.scss
@@ -1,49 +1,0 @@
-.image-container {
-  // animation: slideIn 0.5s ease-in, fadeIn 0.5s ease-in;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: relative;
-
-  .image {
-    width: 100%;
-    height: 100%;
-    transition: 0.1s;
-    object-fit: cover;
-  }
-}
-
-.image-container:hover .image {
-  height: 95%;
-}
-
-.post-info {
-  width: 100%;
-  height: 95%;
-  opacity: 0;
-  transition: 0.2s ease;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  pointer-events: none;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  color: white;
-  font-weight: 700;
-  font-size: 2.5rem;
-}
-
-.image-container:hover .post-info {
-  background-color: rgba(0,0,0,0.2);
-  opacity: 1;
-  text-align: left;
-  text-shadow: 1px 1px 2px rgba(0,0,0,0.1);
-}
-
-@media screen and (max-width: 991px) {
-  .post-info {
-    font-size: 2rem;
-  }
-}
