+++ src/components/post/_post-preview.scss
@@ -1,0 +1,52 @@
+.post-preview {
+  &__link {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    position: relative;
+  }
+
+  &__link:hover {
+    .post-preview__video,
+    .post-preview__image {
+      height: 95%;
+    }
+
+    .post-preview__info {
+      background-color: rgba(0, 0, 0, 0.2);
+      opacity: 1;
+      text-align: left;
+      text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.1);
+    }
+  }
+
+  &__image,
+  &__video {
+    width: 100%;
+    height: 100%;
+    transition: 0.1s;
+    object-fit: cover;
+  }
+
+  &__info {
+    width: 100%;
+    height: 95%;
+    opacity: 0;
+    transition: 0.2s ease;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    pointer-events: none;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    color: white;
+    font-weight: 700;
+    font-size: 2.5rem;
+
+    @media (max-width: 991px) {
+      font-size: 2rem;
+    }
+  }
+}
